This release is a collection of fixes to get Crashmail running on
modern 64-bit Linux.  This is the first release in which the
following tools all pass the test suite:

- crashmail (areafix)
- crashstats
- crashlist
- crashgetnode

Use of the legacy ulong/ushort types have been replaced by
the stdint.h uint32_t/uint16_t types throughout the code.

+++ cmnllib/cmnllib.c
@@ -11,11 +11,11 @@
 
 struct idx
 {
-	ushort zone,net,node,point,region,hub;
-	ulong offset;
+	uint16_t zone,net,node,point,region,hub;
+	uint32_t offset;
 };
                   
-ulong cmnlerr;						                                               
+uint32_t cmnlerr;						                                               
 
 uchar *cmnlerrstr[] = 
 {
@@ -27,12 +27,12 @@ uchar *cmnlerrstr[] =
 	"Failed to open nodelist"
 };	
 
-ushort cmnlgetuword(uchar *buf,ulong offset)
+uint16_t cmnlgetuword(uchar *buf,uint32_t offset)
 {
-   return (ushort)(buf[offset]+256*buf[offset+1]);
+   return (uint16_t)(buf[offset]+256*buf[offset+1]);
 }
 
-long cmnlgetlong(uchar *buf,ulong offset)
+long cmnlgetlong(uchar *buf,uint32_t offset)
 {
    return (long) buf[offset]+
 				 	  buf[offset+1]*256+
@@ -71,7 +71,7 @@ void cmnlCloseNL(osFile nl)
 	osClose(nl);
 }
 
-bool cmnlFindNL(osFile nl,uchar *dir,struct cmnlIdx *cmnlidx,uchar *line,ulong len)
+bool cmnlFindNL(osFile nl,uchar *dir,struct cmnlIdx *cmnlidx,uchar *line,uint32_t len)
 {
 	uchar buf[200];
 	uchar nlname[100];
@@ -79,18 +79,18 @@ bool cmnlFindNL(osFile nl,uchar *dir,struct cmnlIdx *cmnlidx,uchar *line,ulong l
 	bool found;
 	osFile fh;
 	uchar binbuf[16];
-	
+
 	osSeek(nl,4,OFFSET_BEGINNING);
-	
+
 	found=FALSE;
-	
+
 	while(!found)
 	{
-	   if(osRead(nl,nlname,100)!=100)
-   	{
-	   	cmnlerr=CMNLERR_NODE_NOT_FOUND;
-		   return(FALSE);
-   	}
+		if(osRead(nl,nlname,100)!=100)
+		{
+			cmnlerr=CMNLERR_NODE_NOT_FOUND;
+			return(FALSE);
+		}
 
 		idx.offset=0;
 
@@ -112,7 +112,7 @@ bool cmnlFindNL(osFile nl,uchar *dir,struct cmnlIdx *cmnlidx,uchar *line,ulong l
 			idx.offset=cmnlgetlong(binbuf,12);
 
 			found=TRUE;
-			
+
 			if(cmnlidx->zone  != idx.zone)  found=FALSE;
 			if(cmnlidx->net   != idx.net)   found=FALSE;
 			if(cmnlidx->node  != idx.node)  found=FALSE;
@@ -127,7 +127,7 @@ bool cmnlFindNL(osFile nl,uchar *dir,struct cmnlIdx *cmnlidx,uchar *line,ulong l
 	{
 		return(TRUE);
 	}
-		
+
 	MakeFullPath(dir,nlname,buf,200);
 
 	if(!(fh=osOpen(buf,MODE_OLDFILE)))
@@ -135,11 +135,11 @@ bool cmnlFindNL(osFile nl,uchar *dir,struct cmnlIdx *cmnlidx,uchar *line,ulong l
 		cmnlerr=CMNLERR_NO_NODELIST;
 		return(FALSE);
 	}
-	
+
 	osSeek(fh,idx.offset,OFFSET_BEGINNING);
 	osFGets(fh,line,len);	
 	osClose(fh);
-	
+
 	return(TRUE);	
 }
