+++ crashmail/mb.h
@@ -14,7 +14,7 @@ struct Messagebase
    bool (*afterfunc)(bool success);
    bool (*importfunc)(struct MemMessage *mm,struct Area *area);
    bool (*exportfunc)(struct Area *area,bool (*handlefunc)(struct MemMessage *mm));
-   bool (*rescanfunc)(struct Area *area,ulong max,bool (*handlefunc)(struct MemMessage *mm));
+   bool (*rescanfunc)(struct Area *area,uint32_t max,bool (*handlefunc)(struct MemMessage *mm));
 };
 
 #endif
